feat: add setting eternal attributes
