+++ tests/end_to_end.rs
@@ -388,41 +388,3 @@ fn test_can_create_destination() -> zip::result::ZipResult<()> {
     assert!(dest.path().join("mimetype").exists());
     Ok(())
 }
-
-#[test]
-fn unix_mode_dos() {
-    use std::io::Cursor;
-    use zip::ZipArchive;
-    // https://github.com/zip-rs/zip2/issues/737
-    // unix_mode_0x081A4000.zip
-    let data = [
-        0x50, 0x4b, 0x03, 0x04, 0x14, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x21, 0x00, 0x85,
-        0x11, 0x4a, 0x0d, 0x0b, 0x00, 0x00, 0x00, 0x0b, 0x00, 0x00, 0x00, 0x0d, 0x00, 0x00, 0x00,
-        0x74, 0x65, 0x73, 0x74, 0x5f, 0x66, 0x69, 0x6c, 0x65, 0x2e, 0x74, 0x78, 0x74, 0x68, 0x65,
-        0x6c, 0x6c, 0x6f, 0x20, 0x77, 0x6f, 0x72, 0x6c, 0x64, 0x50, 0x4b, 0x01, 0x02, 0x2d, 0x00,
-        0x14, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x21, 0x00, 0x85, 0x11, 0x4a, 0x0d, 0x0b,
-        0x00, 0x00, 0x00, 0x0b, 0x00, 0x00, 0x00, // file name length
-        0x0d, 0x00, // extra field length
-        0x00, 0x00, // file comment length
-        0x00, 0x00, // disk number start
-        0x00, 0x00, // internal file attributes
-        0x00, 0x00, // external file attributes
-        0x00, 0x40, 0x1a, 0x08, // relative offset of local header
-        0x00, 0x00, 0x00, 0x00, // filename: test_file.txt
-        0x74, 0x65, 0x73, 0x74, 0x5f, 0x66, 0x69, 0x6c, 0x65, 0x2e, 0x74, 0x78, 0x74,
-        // end central header
-        0x50, 0x4b, 0x05, 0x06, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x01, 0x00, 0x3b, 0x00, 0x00,
-        0x00, 0x36, 0x00, 0x00, 0x00, 0x00, 0x00,
-    ];
-    let mut archive = ZipArchive::new(Cursor::new(data)).unwrap();
-    for i in 0..archive.len() {
-        let file = archive.by_index(i).unwrap();
-        assert_eq!(
-            file.unix_mode(),
-            Some(33204),
-            "file {:?} has unix_mode {:?}",
-            file.name(),
-            file.unix_mode()
-        );
-    }
-}
